refactor: Move Income Toggle component in one file
